combining transaction summary into util function
